+++ CHANGELOG.md
@@ -10,13 +10,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Actions` - Added a release step for the project's pipeline.
+- `GraphService` - Added a `resetGraph` function to clear the current graph data and state.
 
 ### Changed
+
+- `Algorithms` - Adjusted graph generation algorithms to use `resetGraph` instead of `clearGraph` for better state management.
 
 ### Fixed
 
 - `Logger` - Fixed logs being stored in production.
 - `Actions` - Fixed deployment failing but deploying anyway.
+
+### Removed
+
+- `EdgeService` - Removed dead property `numEdges` from the graph's data.
 
 ## [v2.3.0](https://github.com/Synogun/GraphVM/compare/v2.0.0...v2.3.0)